move PR coverage report fetcher to fetchers file
